refactor: chart 컴포넌트 분리
